drone/requirements.txt
pybullet==3.2.5
numpy==1.24.3
requests==2.31.0
paho-mqtt==1.6.1
matplotlib==3.7.2
scipy==1.11.1
pyyaml==6.0.1

